+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+**Features:**
+* Allow writing back components of the selected Address to the form
+
 **Fixes:**
 * Fix error when geocoding and selection is empty